+++ openai_realtime_voice_agent/root/run.sh
@@ -35,6 +35,7 @@ LONGLIVED_TOKEN=$(bashio::config 'longlived_token')
 
 # --- 🤖 Dex ---
 DEX_ADAPTER_URL=$(bashio::config 'dex_adapter_url')
+DEX_ADAPTER_TOKEN=$(bashio::config 'dex_adapter_token')
 
 # --- ⚙️ Advanced ---
 WEBSOCKET_PORT=$(bashio::config 'websocket_port')
@@ -73,6 +74,7 @@ export LONGLIVED_TOKEN
 export HA_DEVICE_ID
 export ASSIST_PIPELINE_NAME
 export DEX_ADAPTER_URL
+export DEX_ADAPTER_TOKEN
 export WEBSOCKET_PORT
 export SESSION_REUSE_TIMEOUT_SECONDS
 export MAX_CONTEXT_MESSAGES